+++ src/win32/controls/textbox.cpp
@@ -98,7 +98,7 @@ void TextBox::readInput()
         return;
 
     auto hwnd = reinterpret_cast<HWND>(handle());
-    auto cb = (size_t)GetWindowTextLengthW(hwnd) + 1;
+    auto cb = (int)GetWindowTextLengthW(hwnd) + 1;
     std::wstring buf(cb, 0);
     GetWindowTextW(hwnd, &buf[0], cb);
     text(toBytes(buf.c_str()));
